Read Shape entity data Split EntityData into ModelEntity and ShapeEntity modules
