+++ src/domain/shared/Color.ts
@@ -1,0 +1,19 @@
+//
+//  Color.ts
+//
+//  Created by Julien Merzoug on 11 Jul 2022.
+//  Copyright 2022 Vircadia contributors.
+//  Copyright 2022 DigiSomni LLC.
+//
+//  Distributed under the Apache License, Version 2.0.
+//  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
+//
+
+// TODO: Type doc? Check Vec3.ts. Create namespace?
+type Color = {
+    red: number | undefined;
+    green: number | undefined;
+    blue: number | undefined;
+};
+
+export type { Color };
